examples: Split code to be shared with other examples in a common file
Coming CVF examples will also deal with socket creation, get
presentation time of packets, etc. This patch moves this code from
current AAF and CRF examples to a `common.c` file, so that it can be
reused by CVF examples.
